Import hotshot only if profiling is enabled.
It was recently discovered that the standard Python profile module comes
with a licence that is not entirely free.  Some Linux distributions
consequently removed it from the standard Python distribution and moved it into
a separate non-free package.  As a result, 'import hotshot.stats' fails
because it tries to reuse the pstats module from the standard profiler.