Add tests for xmlrpc.py
Taken from zope.app.publisher.xmlrpc.

That plus a few other small changes gets test coverage up to 100%.

There's a hack for dealing with XMLRPCResponse objects that have text
instead of bytes as their body that I'm not proud of, but I think the
bug is somewhere in the publisher, not here.